9/14_22:11

+++ src/main/java/com/example/demo/controller/dto/response/UnrealDetail.java
@@ -18,8 +18,8 @@ public class UnrealDetail {
     private Double nowPrice;
     private Double qty;
     private Double remainQty;
-    private Double fee;
-    private Double cost;
+    private Integer fee;
+    private Integer cost;
     private Double marketValue;
     private Double unrealProfit;
     private String profitability;
